requirements.txt
python-telegram-bot==20.5
gspread==5.7.2
oauth2client==4.1.3
requests==2.31.0
pillow==10.0.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
ffmpeg-python